refactor(hub-ui): decompose DashboardApp into three-pane shell
Move GridCard + agent grid/list + role filter into components/agents/
{GridCard,AgentsPage}.tsx (pure movement). DashboardApp becomes the shell
host: keeps modal orchestration + update/CLI-skew banners + toolbar, maps
the old surface state onto PageId, and renders content inside <Shell>. New
library/home pages render a PlaceholderPage (models opens ModelLibrary).
WorkView kept reachable via a transitional page value until its later-phase
removal. No behavior change to existing views.
