+++ mur-hub-gui/ui/src/i18n/zh-TW.ts
@@ -710,4 +710,6 @@ export const zhTW: Table = {
   "nav.mcp": "MCP",
   "nav.models": "模型",
   "nav.plugins": "外掛",
+  "placeholder.body": "此改版即將推出。",
+  "placeholder.open": "開啟模型庫",
 };
